two graphs a row
