+++ src/scenes/Stats/Stats.scss
@@ -1,4 +1,23 @@
-/*
+@import 'colors';
+
 .content-container-stats-scene {
+
+.chart-row {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .chart-column {
+    text-align: center;
+    width: calc((100% - (2 - 1) * 2rem) / 2);
+  }
+
+  .content-block-heading {
+      color: $color-accent;
+    }
+
+
+
 }
-*/
+
